+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Change Log
   https://docs.google.com/document/d/1zj-edWR1ugqhd6ZxC07Rkq6M7I_jqiR-pO598gFg0p8
 * Moved utility scripts view-portal-object and update-portal-object to dcicutils 8.13.3.
 * Ran poetry update.
+* Extensible validators mechanism (see submitr/validators.py) initially for submitted_id.
 
 
 1.1.0